+++ frontend/src/components/google-maps/map/CarFfeineMap.tsx
@@ -6,8 +6,9 @@ import StationMarkersContainer from '@marker/StationMarkersContainer';
 
 import { debounce } from '@utils/debounce';
 import { useExternalValue, useSetExternalState } from '@utils/external-state';
-import { getDisplayPosition, type DisplayPosition } from '@utils/google-maps';
-import { getSessionStorage, setLocalStorage } from '@utils/storage';
+import { getDisplayPosition } from '@utils/google-maps';
+import { isCachedRegion } from '@utils/google-maps/isCachedRegion';
+import { setLocalStorage } from '@utils/storage';
 
 import { getGoogleMapStore } from '@stores/google-maps/googleMapStore';
 import { warningModalActions } from '@stores/layout/warningModalStore';
@@ -28,7 +29,7 @@ import WarningModalContainer from '@ui/WarningModalContainer';
 
 import { INITIAL_ZOOM_SIZE } from '@constants/googleMaps';
 import { QUERY_KEY_STATION_MARKERS } from '@constants/queryKeys';
-import { LOCAL_KEY_LAST_POSITION, SESSION_KEY_LAST_REQUEST_POSITION } from '@constants/storageKeys';
+import { LOCAL_KEY_LAST_POSITION } from '@constants/storageKeys';
 
 const CarFfeineMap = () => {
   return (
@@ -47,44 +48,6 @@ const CarFfeineMap = () => {
       <StationMarkersContainer />
     </>
   );
-};
-
-const getBounds = (displayPosition: DisplayPosition) => {
-  return {
-    northEast: {
-      latitude: displayPosition.latitude + displayPosition.latitudeDelta,
-      longitude: displayPosition.longitude + displayPosition.longitudeDelta,
-    },
-    southWest: {
-      latitude: displayPosition.latitude - displayPosition.latitudeDelta,
-      longitude: displayPosition.longitude - displayPosition.longitudeDelta,
-    },
-  };
-};
-
-const isCachedRegion = (displayPosition: DisplayPosition) => {
-  const cachedDisplayPosition = getSessionStorage<DisplayPosition | null>(
-    SESSION_KEY_LAST_REQUEST_POSITION,
-    null
-  );
-  if (cachedDisplayPosition === null) {
-    return false;
-  }
-  const cachedBounds = getBounds(cachedDisplayPosition);
-  const bounds = getBounds(displayPosition);
-  if (cachedBounds.northEast.latitude < bounds.northEast.latitude) {
-    return false;
-  }
-  if (cachedBounds.northEast.longitude < bounds.northEast.longitude) {
-    return false;
-  }
-  if (cachedBounds.southWest.latitude > bounds.southWest.latitude) {
-    return false;
-  }
-  if (cachedBounds.southWest.longitude > bounds.southWest.longitude) {
-    return false;
-  }
-  return true;
 };
 
 const CarFfeineMapListener = () => {
